hosts/fenchurch: Copy the Grub bootloader to Microsoft path after install
My EFI is incredibly buggy so it magically "forgets" of all other
bootloaders than Microsofts bootloader each reboot. Even if you remove
Microsofts bootloader it will re-appear to be the default bootloader.

With a recent NixOS upgrade I got a new Grub and new Grub modules, but
the module needed for the one I had manually copied to the Microsoft
path years ago didn't get updated and therefore didn't work with these
new modules.

So this will fix this for the future, make sure to always update the
Microsoft bootloader path so it's never out of date again.