hack(hosts/server-sparv): Hack with the systemd timer to not restart whenever
I've had a problem with this timer that it restarted *on every nixos
deployment* regardless if something related to it had changed or not.

And I didn't manage to actually configure the timer to not behave this
way even after looking at other timers that works as I would expect.

I should probably learn systemd timers properly some day. ¯\_(ツ)_/¯